controllers updated for new suppliers class

+++ app/Http/Controllers/ProductsController.php
@@ -4,21 +4,18 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-
-use App\User;
+use App\OrderDetail;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\DB;
+use App\Supplier;
 use App\Product;
 use App\Rules\MustBeUniqueToSupplier;
 
 class ProductsController extends Controller
 {
     /**
-     *  protect all routes for supplier access only execpt for [routes]
-     *  which can be accessed by anyone
+     *  All routes are protected by the auth.supplier middleware in the routes file
      */
-    public function __construct()
-    {
-        $this->middleware(['auth', 'auth.supplier'])->except(['addToCart', 'productBySupplier']);
-    }
 
     /**
      * Show the form for creating a new resource.
@@ -27,6 +24,15 @@ class ProductsController extends Controller
 
     public function create(Request $request)
     {
+
+        // Why wont this work???? Not returning correct supplier!
+        // dd(Supplier::find($id)->get()->pluck('is_approved')[0]);
+
+        $id = auth::id();
+        $is_approved = Supplier::where('id', $id)->get()->pluck('is_approved')[0];
+        // abort if not approved
+        abort_if($is_approved == '0', 403);
+
         $data = [
             'title' => 'Add New Dish',
             'form_type' => 'create' // used to select from type
@@ -40,8 +46,11 @@ class ProductsController extends Controller
      */
     public function store(Request $request, Product $product)
     {
+        $supplier_id = Supplier::where('user_id', Auth::id())->value('id');
+
         $validatedData = $this->validateInputs($request, $product);
-        $validatedData['user_id'] = $request->user_id;
+
+        $validatedData['supplier_id'] = $supplier_id;
 
         // create the product
         $product = Product::create($validatedData);
@@ -58,6 +67,8 @@ class ProductsController extends Controller
      */
     public function edit(Product $product)
     {
+        $this->authorize('update', $product); // abort if not authorised
+
         $data = [
             'title' => 'Edit Product',
             'form_type' => 'edit', // used to select from type
@@ -67,13 +78,14 @@ class ProductsController extends Controller
         return view('products.edit')->with($data);
     }
 
-
     /**
      * Update the specified resource in storage.
      * @param  Product  current $product object
      */
     public function update(Request $request, Product $product)
     {
+
+        $this->authorize('update', $product); // abort if not authorised
 
         // if this product_id = product_id then it is ok to save
         // ignore this id
@@ -84,8 +96,6 @@ class ProductsController extends Controller
                 'image' => 'file|image|max:3000'
             ]);
         };
-
-
 
         // add and store the product image
         $this->storeImage($product);
@@ -102,11 +112,12 @@ class ProductsController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('update', $product); // abort if not authorised
+
         $product->delete();
 
         return back();
     }
-
 
     /**
      * update the product image.
@@ -131,12 +142,16 @@ class ProductsController extends Controller
      */
     public function validateInputs(Request $request, Product $product)
     {
+
+        // [new MustBeUniqueToSupplier($supplier_id, $productTitle)]
+        $supplier_id = Supplier::where('user_id', Auth::id())->value('id');
+
         // if product_id then record exists
         if ($product->id == null) {
             $validatedData = $request->validate([
                 'name' => [
                     'required',
-                    new MustBeUniqueToSupplier(auth()->user(), $request->name)
+                    new MustBeUniqueToSupplier($supplier_id, $request->name)
                 ],
                 'price' => 'required|numeric|min:0',
             ]);
@@ -144,13 +159,11 @@ class ProductsController extends Controller
             $validatedData = $request->validate([
                 'name' => [
                     'required',
-                    new MustBeUniqueToSupplier(auth()->user(), $request->name, $product->id)
+                    // new MustBeUniqueToSupplier($supplier_id, $request->name, $product->id)
                 ],
                 'price' => 'required|numeric|min:0',
             ]);
         }
-
-
 
         // if there is a file validate that it is an image type
         if (request()->hasFile('image')) {
@@ -163,20 +176,29 @@ class ProductsController extends Controller
     }
 
     /**
-     * Display a list of products by supplier
-     * @param integer $id is user_id
+     * return collection of the 5 most popular products sold in the
+     * last 30 days
      */
-    public function productBySupplier($id)
+    public function mostPopular()
     {
 
-        $user = User::find($id);
+        // return collection of orderDetails from the last 30 days
+        $recentSales = OrderDetail::where('created_at', '>', Carbon::now()->subDays(30))->get();
+
+        $products = DB::table('order_details')
+            ->groupBy('product_id')
+            ->select('product_id', DB::raw('sum(qty) as qty'))
+            ->where('created_at', '>', Carbon::now()->subDays(30))
+            ->orderBy('qty', 'desc')
+            ->take(5)
+            ->get();
 
         $data = [
-            'title' => 'Menu',
-            'products' => $user->products()->paginate(4),
-            'supplier' => $user
+            'title' => 'Most Popular Items',
+            'products' => $products
         ];
 
-        return view('products.product-by-supplier')->with($data);
+        // dd($products);
+        return view('products.most-popular')->with($data);
     }
 }
